create demoapp, simply helloworld
