+++ examples/demoapp/README.md
@@ -1,0 +1,18 @@
+## Demo App Spring Boot Example
+
+
+This example shows how to use many of the features of the Spring Boot rule for Bazel.
+
+To build:
+
+```bash
+bazel build //examples/demoapp
+```
+
+To run:
+
+```bash
+bazel run //examples/demoapp
+```
+
+For full documentation, see the [//springboot](../../springboot) package documentation.
