+++ examples/demoapp/demoapp_classpath.idx
