+++ examples/demoapp/demoapp_dupeclass_allowlist.txt
