+++ examples/demoapp/libs/lib1/src/main/java/com/bazel/demo/IntentionalDupedClass.java
@@ -4,17 +4,10 @@
  * Licensed under the BSD 3-Clause license.
  * For full license text, see LICENSE.txt file in the repo root  or https://opensource.org/licenses/BSD-3-Clause
 */
+package com.bazel.demo;
 
-package com.sample;
-
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-public class SampleRest {
-
-    @RequestMapping("/")
+public class IntentionalDupedClass {
     public String hello() {
-        return "Hello!";
+      return "Hello LIB1!";
     }
 }
