+++ examples/demoapp/libs/lib2/src/test/java/com/bazel/demo/IntentionalDupedClassTest.java
@@ -7,7 +7,6 @@
 package com.bazel.demo;
 
 import static org.junit.Assert.assertEquals;
-
 import org.junit.Test;
 
 public class IntentionalDupedClassTest {
