+++ examples/demoapp/src/main/java/com/sample/SampleRest.java
@@ -7,14 +7,13 @@
 
 package com.sample;
 
-import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.RequestMapping;
 
 @RestController
 public class SampleRest {
-
     @RequestMapping("/")
     public String hello() {
-        return "Hello!";
+      return "Hello!";
     }
 }
