+++ examples/demoapp/src/main/resources/application.properties
@@ -1,0 +1,11 @@
+#
+# Copyright (c) 2017-2021, salesforce.com, inc.
+# All rights reserved.
+# Licensed under the BSD 3-Clause license.
+# For full license text, see LICENSE.txt file in the repo root  or https://opensource.org/licenses/BSD-3-Clause
+#
+
+# expose the actuator endpoints
+#  http://localhost:8080/actuator/beans
+#  http://localhost:8080/actuator/info
+management.endpoints.web.exposure.include=*
