+++ examples/demoapp/src/test/java/com/sample/SampleRestFuncTest.java
@@ -6,14 +6,14 @@
 */
 package com.sample;
 
-import static org.junit.Assert.assertEquals;
-
-import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.TestPropertySource;
 import org.springframework.test.context.junit4.SpringRunner;
+
+import static org.junit.Assert.assertEquals;
+import org.junit.Test;
 
 /**
  * This test class starts the Spring Boot application. Because the Spring context is
