+++ examples/demoapp/src/test/java/com/sample/SampleRestUnitTest.java
@@ -7,7 +7,6 @@
 package com.sample;
 
 import static org.junit.Assert.assertEquals;
-
 import org.junit.Test;
 
 /**
