+++ examples/demoapp/src/test/resources/test.properties
@@ -1,0 +1,1 @@
+anyprop=somevalue
