+++ examples/helloworld/libs/lib1/src/test/java/com/bazel/demo/IntentionalDupedClassTest.java
@@ -7,6 +7,7 @@
 package com.bazel.demo;
 
 import static org.junit.Assert.assertEquals;
+
 import org.junit.Test;
 
 public class IntentionalDupedClassTest {
